pyproject.toml
[project]
name = "stock-tel"
version = "0.1.0"
description = "텔레그램 채널 기반 키움증권 자동매매 시스템"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.55.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "rich>=13.0.0",
    "telethon>=1.36.0",
    "streamlit>=1.51.0",
    "plotly>=6.4.0",
    "pandas>=2.3.3",
    "watchdog>=6.0.0",
]
